feat: test new lectures page
